submission-code/src/submission_code/requirements.txt
/nlc2cmd/src/submission_code/lib/frustrate/vec4ir
numpy
attrs
numba